Rename usecase getCampaignByCode
Try to respect more the usecase naming rules,
so rename this usecase in retieveCampaignInformation.
