+++ api/lib/application/campaigns/campaign-controller.js
@@ -31,7 +31,7 @@ module.exports = {
     const userId = requestUtils.extractUserIdFromRequest(request);
     await _validateFilters(filters);
 
-    const campaign = await usecases.getCampaignByCode({ code: filters.code });
+    const campaign = await usecases.retrieveCampaignInformation({ code: filters.code });
     await usecases.assertUserBelongToOrganization({ userId, campaign });
     const campaignWithLogo = await usecases.addOrganizationLogoToCampaign({ campaign });
