+++ api/lib/domain/usecases/index.js
@@ -90,7 +90,6 @@ module.exports = injectDependencies({
   getAssessment: require('./get-assessment'),
   getAttendanceSheet: require('./get-attendance-sheet'),
   getCampaign: require('./get-campaign'),
-  getCampaignByCode: require('./get-campaign-by-code'),
   getCampaignParticipation: require('./get-campaign-participation'),
   getCampaignParticipationResult: require('./get-campaign-participation-result'),
   getCampaignReport: require('./get-campaign-report'),
@@ -125,6 +124,7 @@ module.exports = injectDependencies({
   rememberUserHasSeenNewProfileInfo: require('./remember-user-has-seen-new-profile-info'),
   removeAllCacheEntries: require('./remove-all-cache-entries'),
   resetScorecard: require('./reset-scorecard'),
+  retrieveCampaignInformation: require('./retrieve-campaign-information'),
   retrieveLastOrCreateCertificationCourse: require('./retrieve-last-or-create-certification-course'),
   saveCertificationCenter: require('./save-certification-center'),
   shareCampaignResult: require('./share-campaign-result'),
