+++ api/lib/domain/usecases/retrieve-campaign-information.js
